update
Former-commit-id: 9bcd874cbeab49fdad608069a25c5dfab609b6ca

+++ server/map.go
@@ -64,6 +64,9 @@ func (m *Map) AddObject(obj IMapObject) (string, bool) {
 }
 
 func (m *Map) DeleteObject(obj IMapObject) {
+	if obj.GetID() == 0 {
+		return
+	}
 	coordinate := obj.GetCoordinate()
 	grid := m.AOI.GetGridByCoordinate(coordinate)
 	grid.DeleteObject(obj)
